requirements.txt
requests>=2.31.0
feedparser>=6.0.11
pandas>=2.0.0
tabulate>=0.9.0
python-dateutil>=2.8.2
sqlalchemy>=2.0.0
pydantic>=2.6.0
python-dotenv>=1.0.0
